% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_read_x3p}
\alias{rust_read_x3p}
\title{Read an X3P file into raw components.}
\usage{
rust_read_x3p(path, verify_checksums)
}
\description{
\code{data}/\code{mask} are returned X-fastest, i.e. ordered to fill an \code{nx}-by-\code{ny}
R matrix in column-major order (the \code{x3ptools} convention).
}
